OPS: Move to uv and pyproject.toml (#846)
* ENH: Reduce installed package size

* FIX: Avoid duplicate version definition

* OPS: Ignore common IDEs

* REF: Move benchmarks utils into benchmarks

* REF: Remove coverge script in favour of pytest-cov

* REF: Moved readme to markdown

* OPS: Move to pyproject.toml and uv for package metadata and build management

* FIX: Remove legacy AUTHORS file from manifest

* DOC: Warning admonitions for release workflows

* OPS: Avoid using importlib on every package import

BREAKING-CHANGE: Removed the __version__ and VERSION values from the __init__.py file. Anyone relying on these for direct runtime introspection should use the get_version() function instead.

* DOC: Remove requirements from readme which are self evident from the pyproject.toml

* DOC: Brought installation instructions up to date

* TST: Make local test matrix work with uv and tox, update docs

* WIP: Progress on test workflow

* OPS: Setting up test matrix for gh actions

* FIX: Correct ends

* OPS: Make tox output verbose

* OPS: Add dev dependency for testing doc build

* OPS: Add other docs dependencies

* OPS: Parameteris gh pythonenvlist

* OPS: Remove no-longer-used requirements file

* OPS: Remove no-longer-used requirements file

* DOC: Add note about django unfold
