+++ .github/workflows/release-publish.yml
@@ -25,19 +25,22 @@ jobs:
         with:
           ref: ${{ inputs.ref }}
 
+      - name: Install uv
+        uses: astral-sh/setup-uv@v5
+
       - name: Install python
         uses: actions/setup-python@v5
         with:
-          python-version: "3.12"
+          # Installs the latest version compatible with the requires-python field
+          python-version-file: "pyproject.toml"
 
-      - name: Install `setuptools`
-        # TODO Remove this legacy step by shifting to pyproject.toml and using a better packaging tool like uv
-        run: python -m pip install --upgrade setuptools
+      - name: Install the project
+        run: uv sync --all-extras --dev
 
       - name: Get package version
         id: version
         run: |
-          VERSION=$(python setup.py --version)
+          VERSION=$(uvx --from=toml-cli toml get --toml-path=pyproject.toml project.version)
           echo "package_version=$VERSION" >> $GITHUB_OUTPUT
 
       - name: Determine if version is a candidate (prerelease)
@@ -92,16 +95,17 @@ jobs:
         with:
           ref: ${{ inputs.ref }}
 
+      - name: Install uv
+        uses: astral-sh/setup-uv@v5
+
       - name: Install python
         uses: actions/setup-python@v5
         with:
-          python-version: "3.12"
+          # Installs the latest version compatible with the requires-python field
+          python-version-file: "pyproject.toml"
 
-      - name: Install `build`
-        run: python -m pip install --upgrade build
-
-      - name: Build package
-        run: python -m build
+      - name: Install the project
+        run: uv build
 
       - name: Publish package distributions to PyPI
         uses: pypa/gh-action-pypi-publish@v1.12.4
