+++ .readthedocs.yaml
@@ -9,12 +9,11 @@ build:
   os: ubuntu-22.04
   tools:
     python: "3.12"
+  jobs:
+    # See https://github.com/astral-sh/uv/issues/10074
+    post_install:
+      - pip install uv
+      - UV_PROJECT_ENVIRONMENT=$READTHEDOCS_VIRTUALENV_PATH uv sync --all-extras --group dev --link-mode=copy
 
 mkdocs:
   configuration: mkdocs.yml
-
-# Declare the Python requirements required to build the documentation
-# See https://docs.readthedocs.io/en/stable/guides/reproducible-builds.html
-python:
-  install:
-    - requirements: docs/requirements.txt
