--- AUTHORS
@@ -1,67 +1,0 @@
-Authors ordered by first contribution
-
-- Lukasz Balcerzak <lukaszbalcerzak@gmail.com>
-- Cesar Canassa <cesar.canassa@gmail.com>
-- Vincent Driessen <vincent@datafox.nl>
-- John Hensley <john@fairviewcomputing.com>
-- Ramanan Sivaranjan <ramanan@funkaoshi.com>
-- Woosuk Suh <pipoket88@gmail.com>
-- Bojan Mihelac <bmihelac@mihelac.org>
-- Rafael Ponieman <rafadev@gmail.com>
-- Daniel Sokolowski <daniel.sokolowski@danols.com>
-- Ali Lozano <alilozanoc@gmail.com>
-- BJ Dierkes <derks@bjdierkes.com>
-- Rach Belaid <rachid.belaid@gmail.com>
-- Michael Crosby <crosby.michael@gmail.com>
-- Greg Hinch <greg@greghinch.com>
-- Aram Dulyan <aram@dulyan.com>
-- Florian Hahn <flo@fhahn.com>
-- Piotr Kilczuk <piotr@tymaszweb.pl>
-- Reavis Sutphin-Gray <reavis@sutphin-gray.com>
-- Adrián López <adrianlopezcalvo@hotmail.com>
-- Jameel Al-Aziz <me@jalaziz.net>
-- John P. Neumann <arrantsquid@gmail.com>
-- Andreas Madsack <andreas@madflex.de>
-- Ivan Kharlamov <the.paper.men@gmail.com>
-- Miguel de Val-Borro <miguel.deval@gmail.com>
-- Jan Nakladal <mojeto1@gmail.com>
-- Yonel Ceruto <yceruto@abalt.org>
-- Luke Faraone <lfaraone@humbughq.com>
-- John Wegis <john@presencelearning.com>
-- Florentin Sardan <florentin.sardan@gmail.com>
-- Geoff Greer <geoff@greer.fm>
-- Hans Larsen <hans@hansl.ca>
-- Fabio C. Barrionuevo da Luz <bnafta@gmail.com> @luzfcb
-- Tomasz Wsuł <2nickers@gmail.com>
-- Xavier Ordoquy <xordoquy@linovia.com>
-- Joshua Bonnett <joshua.bonnett@gmail.com>
-- Jernej Kos <jernej@kos.mx>
-- Bruno Ribeiro da Silva <bruno@e3c.com.br>
-- Cezar Jenkins <emperorcezar@gmail.com>
-- Warren Volz <warren@warrenvolz.com>
-- Omer Katz <omer.drow@gmail.com>
-- Vishal Lal <vish61@gmail.com>
-- Steven DeMartini <sjd@yelp.com>
-- zauddelig <zauddelig@gmail.com>
-- Remco Wendt <remco.wendt@gmail.com>
-- Kevin London <kevinlondon@gmail.com>
-- Kouhei Maeda <mkouhei@gmail.com>
-- Samuel Sutch <sam@sutch.net>
-- Morgan Aubert <morgan.aubert@zoho.com>
-- Brian May <brian@microcomaustralia.com.au>
-- Troy Grosfield <troy.grosfield@gmail.com>
-- Michael Drescher <kaesemeister@gmail.com>
-- Verena Jaspersen <verena.jaspersen@gmail.com>
-- Bertrand Svetchine <bertrand.svetchine@gmail.com>
-- Frank Wickström <frank@bambuser.com>
-- George Karakostas <gckarakostas@gmail.com>
-- Adam Dobrawy <guardian@jawnosc.tk>
-- Jeff Hackshaw <intrepidevio@gmail.com>
-- Chase Bennett <ch@se.gd>
-- Jonny Arnold <jonny.arnold89@gmail.com>
-- Davis Raymond Muro <davisraymondmuro@gmail.com>
-- Richard de Wit <henk.exe@gmail.com>
-- Pedro Rojas Gavidia <pedrorojas.gavidia@gmail.com> @pedrorojasg
-- Rainshaw Gao <rxg@live.com>
-- Meredith Hoo <meredith.hoo1@gmail.com> @zesameri
-- Columbia University, SIS Development and Application Services
