+++ benchmarks/run_benchmarks.py
@@ -26,7 +26,7 @@ django.setup()
 from benchmarks import settings
 from guardian.shortcuts import assign_perm
 from django.core.exceptions import ImproperlyConfigured
-from utils import show_settings
+from benchmarks.utils import show_settings
 from django.contrib.auth.models import User, Group
 from django.utils.termcolors import colorize
 from benchmarks.models import TestModel
