+++ pyproject.toml
@@ -1,0 +1,155 @@
+[project]
+name = "django-guardian"
+version = "3.0.0rc2"
+description = "Per object permissions for Django"
+keywords = ["django", "permissions", "authorization", "object", "row", "level"]
+readme = "README.md"
+license = { text = "BSD-2-Clause" }
+requires-python = ">=3.9"
+classifiers = [
+    "Development Status :: 5 - Production/Stable",
+    "Intended Audience :: Developers",
+    'Environment :: Web Environment',
+    "Topic :: Software Development :: Libraries :: Python Modules",
+    "License :: OSI Approved :: BSD-3-Clause",
+    'Framework :: Django',
+    'Framework :: Django :: 3.2',
+    'Framework :: Django :: 4.1',
+    'Framework :: Django :: 4.2',
+    'Framework :: Django :: 5.0',
+    'Framework :: Django :: 5.1',
+    'Programming Language :: Python',
+    "Programming Language :: Python :: 3 :: Only",
+    "Programming Language :: Python :: 3.9",
+    "Programming Language :: Python :: 3.10",
+    "Programming Language :: Python :: 3.11",
+    "Programming Language :: Python :: 3.12",
+    "Programming Language :: Python :: 3.13",
+    "Operating System :: OS Independent",
+    'Topic :: Security',
+]
+authors = [
+  # In order of first contribution...
+  # New contributors please add your name at the end :)
+  { name = "Lukasz Balcerzak", email = "lukaszbalcerzak@gmail.com" },
+  { name = "Cesar Canassa", email = "cesar.canassa@gmail.com" },
+  { name = "Vincent Driessen", email = "vincent@datafox.nl" },
+  { name = "John Hensley", email = "john@fairviewcomputing.com" },
+  { name = "Ramanan Sivaranjan", email = "ramanan@funkaoshi.com" },
+  { name = "Woosuk Suh", email = "pipoket88@gmail.com" },
+  { name = "Bojan Mihelac", email = "bmihelac@mihelac.org" },
+  { name = "Rafael Ponieman", email = "rafadev@gmail.com" },
+  { name = "Daniel Sokolowski", email = "daniel.sokolowski@danols.com" },
+  { name = "Ali Lozano", email = "alilozanoc@gmail.com" },
+  { name = "BJ Dierkes", email = "derks@bjdierkes.com" },
+  { name = "Rach Belaid", email = "rachid.belaid@gmail.com" },
+  { name = "Michael Crosby", email = "crosby.michael@gmail.com" },
+  { name = "Greg Hinch", email = "greg@greghinch.com" },
+  { name = "Aram Dulyan", email = "aram@dulyan.com" },
+  { name = "Florian Hahn", email = "flo@fhahn.com" },
+  { name = "Piotr Kilczuk", email = "piotr@tymaszweb.pl" },
+  { name = "Reavis Sutphin-Gray", email = "reavis@sutphin-gray.com" },
+  { name = "Adrián López", email = "adrianlopezcalvo@hotmail.com" },
+  { name = "Jameel Al-Aziz", email = "me@jalaziz.net" },
+  { name = "John P. Neumann", email = "arrantsquid@gmail.com" },
+  { name = "Andreas Madsack", email = "andreas@madflex.de" },
+  { name = "Ivan Kharlamov", email = "the.paper.men@gmail.com" },
+  { name = "Miguel de Val-Borro", email = "miguel.deval@gmail.com" },
+  { name = "Jan Nakladal", email = "mojeto1@gmail.com" },
+  { name = "Yonel Ceruto", email = "yceruto@abalt.org" },
+  { name = "Luke Faraone", email = "lfaraone@humbughq.com" },
+  { name = "John Wegis", email = "john@presencelearning.com" },
+  { name = "Florentin Sardan", email = "florentin.sardan@gmail.com" },
+  { name = "Geoff Greer", email = "geoff@greer.fm" },
+  { name = "Hans Larsen", email = "hans@hansl.ca" },
+  { name = "Fabio C. Barrionuevo da Luz", email = "bnafta@gmail.com" },
+  { name = "Tomasz Wsuł", email = "2nickers@gmail.com" },
+  { name = "Xavier Ordoquy", email = "xordoquy@linovia.com" },
+  { name = "Joshua Bonnett", email = "joshua.bonnett@gmail.com" },
+  { name = "Jernej Kos", email = "jernej@kos.mx" },
+  { name = "Bruno Ribeiro da Silva", email = "bruno@e3c.com.br" },
+  { name = "Cezar Jenkins", email = "emperorcezar@gmail.com" },
+  { name = "Warren Volz", email = "warren@warrenvolz.com" },
+  { name = "Omer Katz", email = "omer.drow@gmail.com" },
+  { name = "Vishal Lal", email = "vish61@gmail.com" },
+  { name = "Steven DeMartini", email = "sjd@yelp.com" },
+  { name = "zauddelig", email = "zauddelig@gmail.com" },
+  { name = "Remco Wendt", email = "remco.wendt@gmail.com" },
+  { name = "Kevin London", email = "kevinlondon@gmail.com" },
+  { name = "Kouhei Maeda", email = "mkouhei@gmail.com" },
+  { name = "Samuel Sutch", email = "sam@sutch.net" },
+  { name = "Morgan Aubert", email = "morgan.aubert@zoho.com" },
+  { name = "Brian May", email = "brian@microcomaustralia.com.au" },
+  { name = "Troy Grosfield", email = "troy.grosfield@gmail.com" },
+  { name = "Michael Drescher", email = "kaesemeister@gmail.com" },
+  { name = "Verena Jaspersen", email = "verena.jaspersen@gmail.com" },
+  { name = "Bertrand Svetchine", email = "bertrand.svetchine@gmail.com" },
+  { name = "Frank Wickström", email = "frank@bambuser.com" },
+  { name = "George Karakostas", email = "gckarakostas@gmail.com" },
+  { name = "Adam Dobrawy", email = "guardian@jawnosc.tk" },
+  { name = "Jeff Hackshaw", email = "intrepidevio@gmail.com" },
+  { name = "Chase Bennett", email = "ch@se.gd" },
+  { name = "Jonny Arnold", email = "jonny.arnold89@gmail.com" },
+  { name = "Davis Raymond Muro", email = "davisraymondmuro@gmail.com" },
+  { name = "Richard de Wit", email = "henk.exe@gmail.com" },
+  { name = "Pedro Rojas Gavidia", email = "pedrorojas.gavidia@gmail.com" },
+  { name = "Rainshaw Gao", email = "rxg@live.com" },
+  {name = "Tom Clark", email = "guardian@octue.com"},
+  { name = "Meredith Hoo", email = "meredith.hoo1@gmail.com" },
+  { name = "Chris Maggiuli" },
+  { name = "Bona Fide IT GmbH"},
+  { name = "Columbia University, SIS Development and Application Services" },
+]
+maintainers = [
+  {name = "Tom Clark", email = "django-guardian@octue.com"},
+  { name = "Brian May", email = "brian@microcomaustralia.com.au" },
+  { name = "Lukasz Balcerzak", email = "lukaszbalcerzak@gmail.com" },
+  { name = "Adam Dobrawy", email = "guardian@jawnosc.tk" },
+  { name = "Michael K." },
+  { name = "Chris Maggiuli" },
+  { name = "Bona Fide IT GmbH"},
+]
+dependencies = [
+    "django>=3.2",
+]
+
+[project.urls]
+Homepage = "https://github.com/django-guardian/django-guardian"
+Documentation = "https://django-guardian.readthedocs.io/"
+Repository = "https://github.com/django-guardian/django-guardian"
+Issues = "https://github.com/me/spam/issues"
+Changelog = "https://github.com/django-guardian/django-guardian/releases"
+
+[dependency-groups]
+# PACKAGES=( 'psycopg2-binary' 'mysqlclient')
+dev = [
+    "django-environ>=0.12.0",
+    "mkdocs>=1.6.1",
+    "mkdocs-material>=9.6.12",
+    "mkdocs-material-extensions>=1.3.1",
+    "mkdocstrings>=0.29.1",
+    "mkdocstrings-python>=1.16.10",
+    "mock>=5.1.0",
+    "mysqlclient>=2.2.7",
+    "psycopg2-binary>=2.9.10",
+    "pytest>=8.3.4",
+    "pytest-cov>=5.0.0",
+    "pytest-django>=4.9.0",
+]
+
+[tool.setuptools]
+packages = [
+    "guardian",
+    "guardian.conf",
+    "guardian.management",
+    "guardian.management.commands",
+    "guardian.migrations",
+    "guardian.templatetags",
+    "guardian.testapp",
+    "guardian.testapp.migrations",
+    "guardian.testapp.tests"
+]
+
+[build-system]
+requires = ["setuptools>=61", "wheel"]
+build-backend = "setuptools.build_meta"
