--- setup.cfg
@@ -1,24 +1,0 @@
-[bumpversion]
-current_version = 3.0.0rc1
-
-[build_sphinx]
-source-dir = docs/
-build-dir = docs/build
-all_files = 1
-
-[upload_sphinx]
-upload-dir = docs/build/html
-
-[bdist_rpm]
-requires = Django >= 3.2
-
-[aliases]
-test = pytest
-
-[bumpversion:file:setup.py]
-search = version = '{current_version}'
-replace = version = '{new_version}'
-
-[bumpversion:file:guardian/__init__.py]
-search = __version__ = '{current_version}'
-replace = __version__ = '{new_version}'
