+++ tox.ini
@@ -1,23 +1,41 @@
 [tox]
-envlist = # sort by django version, next by python version
-    {core,example,docs}-py{39,310}-django32,
-    {core,example,docs}-py{39,310,311}-django41,
-    {core,example,docs}-py{39,310,311}-django42,
-    {core,example,docs}-py{310,311,312}-django50,
-    {core,example,docs}-py{310,311,312,313}-django51,
-    {core,example,docs}-py{310,311,312,313}-djangomain,
+envlist =
+    core-py{39,310}-{django32},
+    core-py{39,310,311}-{django41,django42},
+    core-py{310,311,312}-{django50},
+    core-py{310,311,312,313}-{django51},
+    core-py{312,313}-{djangomain},
+    {example,examplegroup}-{py313}-{django51},
+    docs-{py313}-{django51}
+
+
+[gh]
+python =
+    3.9 = core-py39-{django32,django41,django42}
+    3.10 = core-py310-{django32,django41,django42,django50,django51}
+    3.11 = core-py311-{django41,django42,django50,django51}
+    3.12 = core-py312-{django50,django51,djangomain}
+    3.13 = core-py313-{django51,djangomain,django51}, examplegroup-py313-django51, docs-py313-django51
+
 
 [testenv]
-passenv = DATABASE_URL
+recreate = true
+
+allowlist_externals = pytest,mkdocs
+
+pass_env = ["DATABASE_URL", "DIFF_AGAINST", "PYTEST_*", "TOX_GH_MAJOR_MINOR"]
+
 basepython =
     py39: python3.9
     py310: python3.10
     py311: python3.11
     py312: python3.12
     py313: python3.13
+
 changedir =
     example: example_project
-    docs: docs
+    examplegroup: example_project_custom_group
+
 commands =
     django32: python {toxinidir}/manage.py makemigrations --check --dry-run
     django41: python {toxinidir}/manage.py makemigrations --check --dry-run
@@ -25,23 +43,18 @@ commands =
     django50: python {toxinidir}/manage.py makemigrations --check --dry-run
     django51: python {toxinidir}/manage.py makemigrations --check --dry-run
     djangomain: python {toxinidir}/manage.py makemigrations --check --dry-run
-    core: py.test --cov=guardian
-    docs: sphinx-build -b html -d {envtmpdir}/doctrees . {envtmpdir}/html
+    core: pytest --cov=guardian --cov-report=xml
+    docs: mkdocs build
     example: python manage.py test
+    examplegroup: python manage.py test
+
+dependency_groups = 
+  dev
+  
 deps =
-    django-environ
-    core: mock>=0.7.2
-    core: setuptools>=17.1
-    core: pytest
-    core: pytest-django
-    core: pytest-cov
-    example: .
-    docs: sphinx
-    docs: sphinx_rtd_theme
-    docs: setuptools_scm
-    django32: Django>=3.2,<3.3
-    django41: Django>=4.1,<4.2
-    django42: Django>=4.2,<4.3
-    django50: Django>=5.0,<5.1
-    django51: Django>=5.1,<5.2
+    django32: django>=3.2,<4.0
+    django41: django>=4.1,<4.2
+    django42: django>=4.2,<4.3
+    django50: django>=5.0,<5.1
+    django51: django>=5.1,<5.2
     djangomain: https://github.com/django/django/archive/main.tar.gz
